Create ticket.c
The above program prints the availability of the numbers of seats in the bus and it also inform the passengers the arrival time,departure time. it allows the customers to reserve seat and also there is an option for them to cancel the reservation as well.in addition to this the passengers have the option to choose whichever seat they want  based on the available seats.